refactor: renomeia diretorio
